refactor(analysis): split grid owner modules
