+++ docs/architecture-consolidation-plan.md
@@ -61,6 +61,7 @@ fixtures:
 | --- | --- | ---: | --- |
 | Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 1451; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; remaining work is mostly test split and aggregation cleanup |
 | Analysis orchestration | `crates/tokmd-analysis/src/analysis/mod.rs` and `crates/tokmd-analysis/src/analysis/` | coordinator 152; setup 20; files 53; outputs 22; enricher owners <=160 | Derived/source setup, file collection, optional output storage, and inventory/content/git/semantic/code-quality/effort enrichers now live in owner modules while `analyze` remains the public coordinator. Keep `analysis_orchestration` proof scoped to both the historical `analysis.rs` path and the new `analysis/**` owner paths during the file-to-folder transition |
+| Analysis preset grid | `crates/tokmd-analysis/src/grid/mod.rs`, `grid/presets.rs`, and `grid/disabled_feature.rs` | `mod.rs` 10; preset matrix owner 445; disabled-feature warning owner 66 | Keep `mod.rs` as the public barrel; keep preset identity, preset-to-enricher matrix, and plan lookup in `presets.rs`; keep disabled-feature warning text in `disabled_feature.rs`. Preserve the `analysis_orchestration` proof scope. |
 | Analysis API surface | `crates/tokmd-analysis/src/api_surface/mod.rs` and `crates/tokmd-analysis/src/api_surface/` | `mod.rs` 13; report owner 227; symbol dispatcher 56; language scanners <=68; symbol tests 385 | Keep `mod.rs` as a thin coordinator, report aggregation in `report.rs`, source scanning in language owner modules under `symbols`, and leave large integration tests under `api_surface/tests` |
 | Analysis effort implementation | `crates/tokmd-analysis/src/effort/size_basis.rs` and `crates/tokmd-analysis/src/effort/classify.rs` | size-basis aggregation 271; file classification owner 298 | Authored/generated/vendored line aggregation stays in `size_basis.rs`; `.gitattributes` parsing, generated sentinels, path-pattern matching, and heuristic file tagging live in the classification owner. Keep this under the `analysis_effort` proof scope. |
 | Context packing and handoff | `crates/tokmd/src/context_pack.rs`, `crates/tokmd/src/context_pack/`, `crates/tokmd/src/commands/context.rs`, and `crates/tokmd/src/commands/handoff.rs` | `context_pack.rs` 15; selection coordinator 305 + policy owner/tests 192 + selection tests 1368; ranking/packing owner 165; output/log submodule 224; render submodule 204; manifest submodule 195; budget parser/tests 220; context command 218; handoff command 268; handoff capability owner 113; handoff intelligence coordinator 196 + complexity owner 434; handoff output owner/tests 242 | Budget parsing, file selection, inclusion-policy preparation, selection proof tests, ranking/packing, bundle text rendering, single-output/log writing, context bundle manifest writing, handoff capability detection, handoff intelligence construction, lightweight source-complexity estimation, and handoff artifact output writing now live in owner modules; keep context/handoff proof scoped while future handoff splits depend on fresh evidence |
